requirements.txt
pypdf
langchain
langchain-community
sentence-transformers
ctransformers==0.2.25
accelerate
flask
pinecone-client
-e .
langchain_pinecone
python-dotenv
langchain-huggingface
torch
chainlit